changed to sha 256, used different predicates, moved functionality to a different file
